refactor: unifica pasta hooks, move @types para devDependencies, remove jquery e internacionaliza pagination
