+++ package.json
@@ -12,13 +12,10 @@
         "analyze": "ANALYZE=true next build"
     },
     "dependencies": {
-        "@types/react": "^19.2.14",
-        "@types/react-dom": "^19.2.3",
         "axios": "^1.15.2",
         "i18next": "^26.0.6",
         "i18next-browser-languagedetector": "^8.2.1",
         "i18next-http-backend": "^4.0.0",
-        "jquery": "^4.0.0",
         "next": "^16.2.6",
         "next-i18next": "^16.0.5",
         "primeicons": "^7.0.0",
@@ -38,6 +35,8 @@
         "@testing-library/user-event": "^14.6.1",
         "@types/jest": "^30.0.0",
         "@types/node": "^20",
+        "@types/react": "^19.2.14",
+        "@types/react-dom": "^19.2.3",
         "@types/react-slick": "^0.23.13",
         "eslint": "^10.2.1",
         "eslint-config-next": "^16.2.4",
