+++ public/locales/en-US/common.json
@@ -10,6 +10,9 @@
   "page": "Page",
   "of": "of",
   "items": "items",
+  "showing": "Showing",
+  "itemsPerPage": "Items",
+  "pageOf": "Page {{current}} of {{total}}",
   "showMore": "Show more",
   "showLess": "Show less",
   "close": "Close",
