+++ public/locales/pt-BR/common.json
@@ -10,6 +10,9 @@
   "page": "Página",
   "of": "de",
   "items": "itens",
+  "showing": "Exibindo",
+  "itemsPerPage": "Itens",
+  "pageOf": "Página {{current}} de {{total}}",
   "showMore": "Mostrar mais",
   "showLess": "Mostrar menos",
   "close": "Fechar",
