+++ src/app/tvShows/TvShowsClient.tsx
@@ -4,7 +4,7 @@ import Banner from "../../shared/components/banner/banner";
 import { filterGenres } from "../../shared/utils";
 import { genreApi, tvShows } from "../../shared/api/api";
 import { IGenre, IResponse, IListTvShows } from "../../shared/interfaces";
-import { useFetchData } from "../../shared/hook/useFetchData";
+import { useFetchData } from "../../shared/hooks/useFetchData";
 import "../../styles/home.style.scss";
 import SkeletonBanner from "@/shared/components/skeletonLoading/skeletonBanner";
 import Pagination from "@/shared/components/pagination/pagination";
